+++ packages/data-model/src/database/AcDbDatabase.ts
@@ -31,6 +31,9 @@ import {
   AcGePoint3d,
   AcGePoint3dLike
 } from '@mlightcad/geometry-engine'
+import { AcDbLinetypeTableRecord } from './AcDbLinetypeTableRecord'
+import { AcDbTextStyleTableRecord } from './AcDbTextStyleTableRecord'
+import { AcDbDimStyleTableRecord } from './AcDbDimStyleTableRecord'
 
 /**
  * Event arguments for object events in the dictionary.
@@ -92,7 +95,7 @@ export interface AcDbProgressdEventArgs {
   /** The status of the current sub stage */
   subStageStatus: AcDbStageStatus
   /**
-   * Store data associated with the current sub stage. Its meaning of different sub stages 
+   * Store data associated with the current sub stage. Its meaning of different sub stages
    * are as follows.
    * - 'PARSE' stage: statistics of parsing task
    * - 'FONT' stage: fonts needed by this drawing
@@ -774,6 +777,88 @@ export class AcDbDatabase extends AcDbObject {
       stage: 'FETCH_FILE',
       subStageStatus: 'END'
     })
+  }
+
+  /**
+   * Create default layer, line type, dimension type, text style and layout.
+   */
+  createDefaultData() {
+    // Create default layer
+    const defaultColor = new AcCmColor()
+    defaultColor.colorIndex = 7 // white
+    this._tables.layerTable.add(
+      new AcDbLayerTableRecord({
+        name: '0',
+        standardFlags: 0,
+        linetype: 'Continuous',
+        lineWeight: 0,
+        isOff: false,
+        color: defaultColor,
+        isPlottable: true
+      })
+    )
+
+    // Create default line type
+    this._tables.linetypeTable.add(
+      new AcDbLinetypeTableRecord({
+        name: 'ByBlock',
+        standardFlag: 0,
+        description: '',
+        totalPatternLength: 0
+      })
+    )
+    this._tables.linetypeTable.add(
+      new AcDbLinetypeTableRecord({
+        name: 'ByLayer',
+        standardFlag: 0,
+        description: '',
+        totalPatternLength: 0
+      })
+    )
+    this._tables.linetypeTable.add(
+      new AcDbLinetypeTableRecord({
+        name: 'Continuous',
+        standardFlag: 0,
+        description: 'Solid line',
+        totalPatternLength: 0
+      })
+    )
+
+    // Create default text style
+    this._tables.textStyleTable.add(
+      new AcDbTextStyleTableRecord({
+        name: 'Standard',
+        standardFlag: 0,
+        fixedTextHeight: 0,
+        widthFactor: 1,
+        obliqueAngle: 0,
+        textGenerationFlag: 0,
+        lastHeight: 0.2,
+        font: 'SimKai',
+        bigFont: '',
+        extendedFont: 'SimKai'
+      })
+    )
+
+    // Create default dimension style
+    this._tables.dimStyleTable.add(
+      new AcDbDimStyleTableRecord({
+        name: 'Standard',
+        dimtxsty: 'Standard'
+      })
+    )
+
+    // Create default layer for model space
+    const layout = new AcDbLayout()
+    layout.layoutName = 'Model'
+    layout.tabOrder = 0
+    layout.blockTableRecordId = this._tables.blockTable.modelSpace.objectId
+    layout.limits.min.copy({ x: 0, y: 0 })
+    layout.limits.max.copy({ x: 1000000, y: 1000000 })
+    layout.extents.min.copy({ x: 0, y: 0, z: 0 })
+    layout.extents.max.copy({ x: 1000000, y: 1000000, z: 0 })
+    this._dictionaries.layouts.setAt(layout.layoutName, layout)
+    this._tables.blockTable.modelSpace.layoutId = layout.objectId
   }
 
   /**